Crate slack_api

Low-level, direct interface for the Slack Web API.

Modules

api

auth

bots

channels
Get info on your team's Slack channels, create or archive channels, invite users, set the topic and purpose, and mark a channel as read.

chat
Post chat messages to Slack.

dnd
Adjust and view Do Not Disturb settings for team members.

emoji

files
Get info on files uploaded to Slack, upload new files to Slack.

files_comments

groups
Get info on your team's private channels.

im
Get info on your direct messages.

mpim
Get info on your multiparty direct messages.

oauth

pins

reactions

reminders

requests
Functionality for sending requests to Slack.

rtm

search
Search your team's files and messages.

stars

team

team_profile

usergroups
Get info on your team's User Groups.

usergroups_users

users
Get info on members of your Slack team.

users_profile

Structs

Bot

BotIcons

Channel

ChannelPurpose

ChannelTopic

File

FileComment

Group

GroupPurpose

GroupTopic

Im

MessageBotMessage

MessageBotMessageIcons

MessageChannelArchive

MessageChannelJoin

MessageChannelLeave

MessageChannelName

MessageChannelPurpose

MessageChannelTopic

MessageChannelUnarchive

MessageFileComment

MessageFileMention

MessageFileShare

MessageGroupArchive

MessageGroupJoin

MessageGroupLeave

MessageGroupName

MessageGroupPurpose

MessageGroupTopic

MessageGroupUnarchive

MessageMeMessage

MessageMessageChanged

MessageMessageChangedMessage

MessageMessageChangedMessageEdited

MessageMessageDeleted

MessageMessageReplied

MessageMessageRepliedMessage

MessageMessageRepliedMessageReply

MessagePinnedItem

MessagePinnedItemItem

MessageReplyBroadcast

MessageReplyBroadcastAttachment

MessageStandard

MessageStandardAttachment

MessageStandardAttachmentField

MessageStandardEdited

MessageUnpinnedItem

MessageUnpinnedItemItem

Mpim

Paging

Reaction

Reminder

Team

TeamIcon

ThreadInfo

User

UserProfile

UserProfileFields

Usergroup

UsergroupPrefs

Enums

Message

Functions

default_client
Provides a default reqwest client to give to the API functions to send requests.