Struct slack_api::chat::PostMessageRequest

pub struct PostMessageRequest<'a> {
    pub channel: &'a str,
    pub text: &'a str,
    pub parse: Option<&'a str>,
    pub link_names: Option<bool>,
    pub attachments: Option<&'a str>,
    pub unfurl_links: Option<bool>,
    pub unfurl_media: Option<bool>,
    pub username: Option<&'a str>,
    pub as_user: Option<bool>,
    pub icon_url: Option<&'a str>,
    pub icon_emoji: Option<&'a str>,
    pub thread_ts: Option<&'a str>,
    pub reply_broadcast: Option<bool>,
}

Fields
channel: &'a str

Channel, private group, or IM channel to send message to. Can be an encoded ID, or a name. See below for more details.

text: &'a str

Text of the message to send. See below for an explanation of formatting. This field is usually required, unless you're providing only attachments instead.

parse: Option<&'a str>

Change how messages are treated. Defaults to none. See below.

link_names: Option<bool>

Find and link channel names and usernames.

attachments: Option<&'a str>

Structured message attachments.

unfurl_links: Option<bool>

Pass true to enable unfurling of primarily text-based content.

unfurl_media: Option<bool>

Pass false to disable unfurling of media content.

username: Option<&'a str>

Set your bot's user name. Must be used in conjunction with as_user set to false, otherwise ignored. See authorship below.

as_user: Option<bool>

Pass true to post the message as the authed user, instead of as a bot. Defaults to false. See authorship below.

icon_url: Option<&'a str>

URL to an image to use as the icon for this message. Must be used in conjunction with as_user set to false, otherwise ignored. See authorship below.

icon_emoji: Option<&'a str>

Emoji to use as the icon for this message. Overrides icon_url. Must be used in conjunction with as_user set to false, otherwise ignored. See authorship below.

thread_ts: Option<&'a str>

Provide another message's ts value to make this message a reply. Avoid using a reply's ts value; use its parent instead.

reply_broadcast: Option<bool>

Used in conjunction with thread_ts and indicates whether reply should be made visible to everyone in the channel or conversation. Defaults to false.