Module slack_api::dnd

Adjust and view Do Not Disturb settings for team members.

Structs

EndDndResponse

EndSnoozeResponse

InfoRequest

InfoResponse

SetSnoozeRequest

SetSnoozeResponse

TeamInfoRequest

TeamInfoResponse

Enums

EndDndError

EndSnoozeError

InfoError

SetSnoozeError

TeamInfoError

Functions

end_dnd
Ends the current user's Do Not Disturb session immediately.

end_snooze
Ends the current user's snooze mode immediately.

info
Retrieves a user's current Do Not Disturb status.

set_snooze
Turns on Do Not Disturb mode for the current user, or changes its duration.

team_info
Retrieves the Do Not Disturb status for users on a team.