Enum slack_api::Error

pub enum Error {
    Http(Error),
    JsonDecode(DecoderError),
    JsonParse(ParserError),
    JsonEncode(EncoderError),
    Api(String),
    Internal(String),
}

slack::Error represents errors that can happen while using the RtmClient

Variants

Http

Http client error

JsonDecode

Error decoding Json

JsonParse

Error parsing Json

JsonEncode

Error encoding Json

Api

Slack Api Error

Internal

Errors that do not fit under the other types, Internal is for EG channel errors.