Module slack_api::groups

Get info on your team's private groups.

For more information, see Slack's API documentation.

Structs

AbridgedGroup

ArchiveResponse

CloseResponse

CreateChildResponse

CreateResponse

HistoryResponse

InfoResponse

InviteResponse

KickResponse

LeaveResponse

ListResponse

MarkResponse

OpenResponse

RenameResponse

SetPurposeResponse

SetTopicResponse

UnarchiveResponse

Functions

archive
Archives a private group.

close
Closes a private group.

create
Creates a private group.

create_child
Clones and archives a private group.

history
Fetches history of messages and events from a private group.

info
Gets information about a private group.

invite
Invites a user to a private group.

kick
Removes a user from a private group.

leave
Leaves a private group.

list
Lists private groups that the calling user has access to.

mark
Sets the read cursor in a private group.

open
Opens a private group.

rename
Renames a private group.

set_purpose
Sets the purpose for a private group.

set_topic
Sets the topic for a private group.

unarchive
Unarchives a private group.