Module slack_api::mpim

Get info on your multiparty direct messages.

Structs

CloseRequest

CloseResponse

HistoryRequest

HistoryResponse

ListResponse

MarkRequest

MarkResponse

OpenRequest

OpenResponse

RepliesRequest

RepliesResponse

Enums

CloseError

HistoryError

ListError

MarkError

OpenError

RepliesError

Functions

close
Closes a multiparty direct message channel.

history
Fetches history of messages and events from a multiparty direct message.

list
Lists multiparty direct message channels for the calling user.

mark
Sets the read cursor in a multiparty direct message channel.

open
This method opens a multiparty direct message.

replies
Retrieve a thread of messages posted to a direct message conversation from a multiparty direct message.