Enum slack_api::team::AccessLogsError

pub enum AccessLogsError<E: Error> {
    PaidOnly,
    OverPaginationLimit,
    NotAuthed,
    InvalidAuth,
    AccountInactive,
    UserIsBot,
    InvalidArgName,
    InvalidArrayArg,
    InvalidCharset,
    InvalidFormData,
    InvalidPostType,
    MissingPostType,
    RequestTimeout,
    MalformedResponse(Error),
    Unknown(String),
    Client(E),
}

Variants
PaidOnly

This is only available to paid teams.

OverPaginationLimit

It is not possible to request more than 1000 items per page or more than 100 pages.

NotAuthed

No authentication token provided.

InvalidAuth

Invalid authentication token.

AccountInactive

Authentication token is for a deleted user or team.

UserIsBot

This method cannot be called by a bot user.

InvalidArgName

The method was passed an argument whose name falls outside the bounds of common decency. This includes very long names and names with non-alphanumeric characters other than _. If you get this error, it is typically an indication that you have made a very malformed API call.

InvalidArrayArg

The method was passed a PHP-style array argument (e.g. with a name like foo[7]). These are never valid with the Slack API.

InvalidCharset

The method was called via a POST request, but the charset specified in the Content-Type header was invalid. Valid charset names are: utf-8 iso-8859-1.

InvalidFormData

The method was called via a POST request with Content-Type application/x-www-form-urlencoded or multipart/form-data, but the form data was either missing or syntactically invalid.

InvalidPostType

The method was called via a POST request, but the specified Content-Type was invalid. Valid types are: application/json application/x-www-form-urlencoded multipart/form-data text/plain.

MissingPostType

The method was called via a POST request and included a data payload, but the request did not include a Content-Type header.

RequestTimeout

The method was called via a POST request, but the POST data was either missing or truncated.

MalformedResponse(Error)

The response was not parseable as the expected object

Unknown(String)

The response returned an error that was unknown to the library

Client(E)

The client had an error sending the request to Slack