Trait slack_api::SlackWebRequestSender

pub trait SlackWebRequestSender {
    fn send<'a>(&self, method: &str, params: HashMap<&str, &'a str>) -> Result<String, HttpRequestError>;

    fn send_authed<'a>(&self, method: &str, token: &'a str, params: HashMap<&str, &'a str>) -> Result<String, HttpRequestError> { ... }
}

Functionality for sending authenticated and unauthenticated requests to Slack via HTTP.

Should be implemented for clients to send requests to Slack.

Required Methods

fn send<'a>(&self, method: &str, params: HashMap<&str, &'a str>) -> Result<String, HttpRequestError>

Make an API call to Slack. Takes a map of parameters that get appended to the request as query params. Returns the response body string after checking it has "ok": true, or an Error

Provided Methods

fn send_authed<'a>(&self, method: &str, token: &'a str, params: HashMap<&str, &'a str>) -> Result<String, HttpRequestError>

Make an API call to Slack that includes the configured token. Takes a map of parameters that get appended to the request as query params. Returns the response body string after checking it has "ok": true, or an Error