Crate slack_api

Low-level, direct interface for the Slack Web API.

Modules

api
Checks API calling code.

auth
Checks authentication & identity.

channels
Get info on your team's Slack channels, create or archive channels, invite users, set the topic and purpose, and mark a channel as read.

chat
Post chat messages to Slack.

emoji
For more information, see Slack's API documentation.

files
Get info on files uploaded to Slack, upload new files to Slack.

groups
Get info on your team's private groups.

im
Get info on your direct messages.

oauth
For more information, see Slack's API documentation.

pins
For more information, see Slack's API documentation.

reactions
For more information, see Slack's API documentation.

rtm
For more information, see Slack's API documentation.

search
Search your team's files and messages.

stars
For more information, see Slack's API documentation.

team
For more information, see Slack's API documentation.

users
Get info on members of your Slack team.

Structs

Attachment
The Slack Attachment object as found in richly-formatted messages.

AttachmentField
A field within an Attachment.

Channel
The Slack Channel type.

Comment
The Comment object as found in the File type and the files.info response.

File
The Slack File type.

Group
The Slack Group type.

Im
The Slack Im type.

Pagination
The Paging object as found in API endpoints that return pages of items.

Purpose
The Purpose object as found in the Group and Channel types.

Reaction
The Reaction object as found in the File type and some Message responses.

Team
The Team object as found in the rtm.start response.

Topic
The Topic object as found in the Group and Channel types.

User
The Slack User type.

UserProfile
The Profile that belongs to a User.

Enums

Error
slack::Error represents errors that can happen while using the RtmClient

Event
Represents Slack rtm event types.

Item
Represents a Message, File or Comment as returned by pins.list, reactions.list, and reactions.get.

Message
Represents Slack message event types.

StarredItem
Represents a starred item as returned by stars.list.

Type Definitions

ApiResult