Module slack_api::groups

Get info on your team's private channels.

Structs

ArchiveRequest

ArchiveResponse

CloseRequest

CloseResponse

CreateChildRequest

CreateChildResponse

CreateRequest

CreateResponse

HistoryRequest

HistoryResponse

InfoRequest

InfoResponse

InviteRequest

InviteResponse

KickRequest

KickResponse

LeaveRequest

LeaveResponse

ListRequest

ListResponse

MarkRequest

MarkResponse

OpenRequest

OpenResponse

RenameRequest

RenameResponse

RenameResponseChannel

RepliesRequest

RepliesResponse

SetPurposeRequest

SetPurposeResponse

SetTopicRequest

SetTopicResponse

UnarchiveRequest

UnarchiveResponse

Enums

ArchiveError

CloseError

CreateChildError

CreateError

HistoryError

InfoError

InviteError

KickError

LeaveError

ListError

MarkError

OpenError

RenameError

RepliesError

SetPurposeError

SetTopicError

UnarchiveError

Functions

archive
Archives a private channel.

close
Closes a private channel.

create
Creates a private channel.

create_child
Clones and archives a private channel.

history
Fetches history of messages and events from a private channel.

info
Gets information about a private channel.

invite
Invites a user to a private channel.

kick
Removes a user from a private channel.

leave
Leaves a private channel.

list
Lists private channels that the calling user has access to.

mark
Sets the read cursor in a private channel.

open
Opens a private channel.

rename
Renames a private channel.

replies
Retrieve a thread of messages posted to a private channel

set_purpose
Sets the purpose for a private channel.

set_topic
Sets the topic for a private channel.

unarchive
Unarchives a private channel.